requirements.txt
asyncpraw==7.7.1
python-dotenv==1.0.0  
praw==7.8.1
